+++ practice3/src/buildings/exceptions/InvalidSpaceAreaException.java
@@ -1,7 +1,7 @@
 package buildings.exceptions;
 
 public class InvalidSpaceAreaException extends IllegalArgumentException{
-    public InvalidSpaceAreaException(int space){
+    public InvalidSpaceAreaException(double space){
         String string= "InvalidSpaceAreaException\nInvalid argument: " + space + "\nEnter space between 20 and 500";
         throw new IllegalArgumentException(string);
     }
