+++ practice3/src/buildings/exceptions/SpaceIndexOutOfBoundsException.java
@@ -1,7 +1,7 @@
 package buildings.exceptions;
 
 public class SpaceIndexOutOfBoundsException extends IndexOutOfBoundsException{
-    public SpaceIndexOutOfBoundsException(int num, int size){
+    public SpaceIndexOutOfBoundsException(int num, double size){
         String string = "SpaceIndexOutOfBoundsException\nInvalid argument: " + num + "\nEnter number between 0 and " + (size-1);
         throw new IndexOutOfBoundsException(string);
     }
