+++ practice3/src/buildings/office/OfficeFloor.java
@@ -48,8 +48,8 @@ public class OfficeFloor implements Floor, Serializable, Cloneable {
         return floor.length();
     }
 
-    public int getSumSquare(){
-        int square = 0;
+    public double getSumSquare(){
+        double square = 0;
         for (int i = 0; i < floor.length(); i++){
             square += floor.getNode(i).getSpace().getSquare();
         }
